refactor: clear duplicate code
