+++ src/components/favourite-button/favourite-button.tsx
@@ -1,25 +1,36 @@
 import React from 'react';
 import classNames from 'classnames';
 
-import { Button, Loader } from '../../components';
+import { useAuth, useFavourite } from '../../hooks';
+import { Comics } from '../../models/comics';
 
+import { Button, Loader } from '../../components';
 import { HeartIcon } from '../icons/heart-icon';
 
 import s from './favourite-button.module.css';
 
 interface Props {
-	isComicsFavourite: boolean;
-	isAuth: boolean;
-	isDisabled: boolean;
-	onFavouriteClick: () => void;
+	comics: Comics;
 }
 
-export function FavouriteButton({
-	isComicsFavourite,
-	isAuth,
-	isDisabled,
-	onFavouriteClick
-}: Props) {
+export function FavouriteButton({ comics }: Props) {
+	const { isAuth } = useAuth();
+	const { isFavourite, addToFavourite, removeFromFavourite, isDisabled } =
+		useFavourite();
+
+	const [isComicsFavourite, setIsComicsFavourite] = React.useState(
+		isFavourite(Number(comics.id))
+	);
+
+	const onFavouriteClick = () => {
+		if (!isComicsFavourite) {
+			addToFavourite(comics);
+		} else {
+			removeFromFavourite(comics);
+		}
+		setIsComicsFavourite(prevState => !prevState);
+	};
+
 	return (
 		<Button
 			className={classNames(s.favourite, {
