+++ src/components/item-card/item-card.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 
 import { FormatEnum, getImage } from '../../utils/images';
-import { useAuth, useFavourite } from '../../hooks';
 import { Comics } from '../../models/comics';
 
 import { FavouriteButton } from '../favourite-button/favourite-button';
@@ -14,23 +13,6 @@ interface Props {
 }
 
 export function ItemCard({ comics }: Props) {
-	const { isAuth } = useAuth();
-	const { isFavourite, addToFavourite, removeFromFavourite, isDisabled } =
-		useFavourite();
-
-	const [isComicsFavourite, setIsComicsFavourite] = React.useState(
-		isFavourite(Number(comics.id))
-	);
-
-	const onFavouriteClick = () => {
-		if (!isComicsFavourite) {
-			addToFavourite(comics);
-		} else {
-			removeFromFavourite(comics);
-		}
-		setIsComicsFavourite(prevState => !prevState);
-	};
-
 	return (
 		<li className={s.card} title={comics.title}>
 			<div className={s.content}>
@@ -49,12 +31,7 @@ export function ItemCard({ comics }: Props) {
 				</Link>
 				<div className={s.info}>
 					<h3 className={s.title}>{comics.title}</h3>
-					<FavouriteButton
-						isAuth={isAuth}
-						isComicsFavourite={isComicsFavourite}
-						isDisabled={isDisabled}
-						onFavouriteClick={onFavouriteClick}
-					/>
+					<FavouriteButton comics={comics} />
 				</div>
 			</div>
 		</li>
