+++ src/pages/comics/comics.tsx
@@ -3,31 +3,14 @@ import { useParams } from 'react-router-dom';
 
 import { useFetchComicsByIdQuery } from '../../store/api/comics-api';
 import { formateDate } from '../../utils/format-date';
-import { useAuth, useFavourite } from '../../hooks';
 import { Loader, FavouriteButton } from '../../components';
 
 import s from './comics.module.css';
 
 function Comics() {
 	const { comicsId } = useParams();
-	const { isAuth } = useAuth();
-	const { isFavourite, addToFavourite, removeFromFavourite, isDisabled } =
-		useFavourite();
 
 	const { data: comics, isLoading } = useFetchComicsByIdQuery(comicsId || '');
-
-	const [isComicsFavourite, setIsComicsFavourite] = React.useState(
-		isFavourite(Number(comicsId))
-	);
-
-	const onFavouriteClick = () => {
-		if (!isComicsFavourite) {
-			addToFavourite(comics);
-		} else {
-			removeFromFavourite(comics);
-		}
-		setIsComicsFavourite(prevState => !prevState);
-	};
 
 	if (isLoading || !comics) {
 		return <Loader />;
@@ -48,12 +31,7 @@ function Comics() {
 				</div>
 				<div className={s.info}>
 					<h2 className={s.title}>{comics.title}</h2>
-					<FavouriteButton
-						isAuth={isAuth}
-						isComicsFavourite={isComicsFavourite}
-						isDisabled={isDisabled}
-						onFavouriteClick={onFavouriteClick}
-					/>
+					<FavouriteButton comics={comics} />
 					<div className={s.data}>
 						<strong>On sale: </strong>
 						<span>{saleDate}</span>
